don't create dummy initrd on innitial install

+++ rpm/post.sh
@@ -25,19 +25,21 @@ fi
 
 touch /lib/modules/%ver_str/modules.dep
 
-if [ -f /etc/fstab ]; then
-    cd /boot
-    /sbin/mkinitrd -k $image -i initrd-%ver_str
+if test "$YAST_IS_RUNNING" != instsys ; then
+	if [ -f /etc/fstab ]; then
+	    cd /boot
+	    /sbin/mkinitrd -k $image -i initrd-%ver_str
 
-    if [ -L /boot/initrd -o ! -e /boot/initrd ]; then
-	if [ -e /boot/initrd-%ver_str ]; then
-	    relink initrd-%ver_str /boot/initrd
-	elif [ -L /boot/initrd ]; then
-	    rm -f /boot/initrd
+	    if [ -L /boot/initrd -o ! -e /boot/initrd ]; then
+		if [ -e /boot/initrd-%ver_str ]; then
+		    relink initrd-%ver_str /boot/initrd
+		elif [ -L /boot/initrd ]; then
+		    rm -f /boot/initrd
+		fi
+	    fi
+	else
+	    echo "please run mkinitrd as soon as your system is complete"
 	fi
-    fi
-else
-    echo "please run mkinitrd as soon as your system is complete"
 fi
 
 # $1 is 1 in postinstall if this package is installed
